refactor(client): modularize chart and indicator logic
This commit refactors the client-side charting implementation to be more modular and maintainable. The monolithic `useChartOptions` hook is broken down into smaller, single-responsibility hooks for chart types, indicators, and layout.

Key client-side changes:
- **Chart Types:** Introduced `useCandlestickSeries`, `useLineSeries`, and `useAreaSeries` hooks.
- **Indicators:** Created dedicated hooks for each indicator (SMA, EMA, Bollinger, RSI, MACD).
- **Layout:** Added `useChartLayout` for dynamic grid and axis management.
- **Calculation:** Centralized indicator calculation logic in `technicalIndicators.ts` and `indicatorProcessor.ts`.
- **Utils:** Added `markLineUtils.ts` for creating price mark lines.

feat(server): add common authentication utilities

This commit also adds a new file `apps/server/src/routes/auth/common.ts` which includes:
- Zod schemas for authentication and user management endpoints.
- Repository instances and Prisma client setup.
- Helper functions for user authentication flows.
